Only run the git-commit if there are changed files
If you try to run git-commit when there are no changes, it returns 1 (an error) even though
it doesn't _feel_ like an error.  So code changed to run git-diff-index (which returns true
IF there is a changed file) first, and then do an AND git-commit.

See source for where I got this idea.